internal: prepare to merge hir::BinaryOp and ast::BinOp
